Split in two project + rename
