--- dockerfiles/base/authorized_keys
@@ -1,1 +1,0 @@
-ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQDJsEwrv4OV9WqP8Y364aKs68bKo1cM6dqXkQfkOnHZ5qHyDzdQOFU2nEr5+IfDVtmWps1qQPh9ReRcS61NLu9Cf4kaq1w+1ISwLbmuxI9MDgBEJVCNXqaa7eT0bxisnVbfPZvcR0AZ/g6pTl1RYmQf5QDW5cdhzRuRfeGyESniCGUbsS6VVXcORK54AWdzPZxhQGQ4uGCLuTQZiybc4wrlu09XsXFY3LtfDTtOmruL/bLkmB2e56Z3CFCDTGOWwQhFLS+Ifj/PVHZBGd9T7Tp8eVD7RzsqTEf/GjP2V4WoD1vNeVLxKxVxbGCI04jb2ReKC6ywwhJXGxmoMmB2kMpP epheo@epheo-HP-Z420-Workstation
