--- dockerfiles/glance/etc/schema-image.json
@@ -1,28 +1,0 @@
-{
-    "kernel_id": {
-        "type": "string",
-        "pattern": "^([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12}$",
-        "description": "ID of image stored in Glance that should be used as the kernel when booting an AMI-style image."
-    },
-    "ramdisk_id": {
-        "type": "string",
-        "pattern": "^([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12}$",
-        "description": "ID of image stored in Glance that should be used as the ramdisk when booting an AMI-style image."
-    },
-    "instance_uuid": {
-        "type": "string",
-        "description": "ID of instance used to create this image."
-    },
-    "architecture": {
-        "description": "Operating system architecture as specified in http://docs.openstack.org/trunk/openstack-compute/admin/content/adding-images.html",
-        "type": "string"
-    },
-    "os_distro": {
-        "description": "Common name of operating system distribution as specified in http://docs.openstack.org/trunk/openstack-compute/admin/content/adding-images.html",
-        "type": "string"
-    },
-    "os_version": {
-        "description": "Operating system version as specified by the distributor",
-        "type": "string"
-    }
-}
