--- dockerfiles/glance/supervisord.conf
@@ -1,15 +1,0 @@
-[supervisord]
-nodaemon=true
-
-[program:sshd]
-command=/usr/sbin/sshd -D
-
-[program:glance-registry]
-command=/usr/bin/glance-registry
-autorestart=true
-priority=30
-
-[program:glance-api]
-command=/usr/bin/glance-api
-autorestart=true
-priority=30
