--- dockerfiles/horizon/supervisord.conf
@@ -1,15 +1,0 @@
-[supervisord]
-nodaemon=true
-
-[program:sshd]
-command=/usr/sbin/sshd -D
-
-[program:httpd]
-command=/bin/bash -c "exec /usr/sbin/httpd -DFOREGROUND"
-autorestart=true
-priority=30
-
-[program:memcached]
-command=/usr/bin/memcached -u daemon
-autorestart=true
-priority=30
