--- dockerfiles/keystone/supervisord.conf
@@ -1,10 +1,0 @@
-[supervisord]
-nodaemon=true
-
-[program:sshd]
-command=/usr/sbin/sshd -D
-
-[program:keystone-all]
-command=/usr/bin/keystone-all
-autorestart=true
-priority=30
