--- dockerfiles/mysql/my.cnf
@@ -1,2 +1,0 @@
-[mysqld]
-bind-address=0.0.0.0
