--- dockerfiles/mysql/mysqld_charset.cnf
@@ -1,7 +1,0 @@
-[mysqld]
-character_set_server=utf8
-character_set_filesystem=utf8
-collation-server=utf8_general_ci
-init-connect='SET NAMES utf8'
-init_connect='SET collation_connection = utf8_general_ci'
-skip-character-set-client-handshake
