--- dockerfiles/nova/Dockerfile
@@ -1,16 +1,0 @@
-FROM octopenstack/oos-base:latest
-MAINTAINER Thibaut Lapierre <root@epheo.eu>
-# http://docs.openstack.org/icehouse/install-guide/install/yum/content/nova-controller.html
-
-# Install Nova
-RUN yum -y install openstack-nova-api openstack-nova-cert openstack-nova-conductor openstack-nova-console openstack-nova-novncproxy openstack-nova-scheduler python-novaclient
-
-#Setup supervisord
-ADD supervisord.conf /etc/supervisord.conf
-
-VOLUME  ["/var/log/nova"]
-
-EXPOSE 9774
-EXPOSE 8775
-
-CMD ["/usr/bin/supervisord", "-n"]
