--- dockerfiles/nova/etc/release
@@ -1,4 +1,0 @@
-[Nova]
-vendor = RDO Project
-product = OpenStack Nova
-package = 2.el7.centos
