--- dockerfiles/nova/supervisord.conf
@@ -1,35 +1,0 @@
-[supervisord]
-nodaemon=true
-
-[program:sshd]
-command=/usr/sbin/sshd -D
-
-[program:nova-api]
-command=/usr/bin/nova-api
-autorestart=true
-priority=30
-
-[program:nova-cert]
-command=/usr/bin/nova-cert
-autorestart=true
-priority=30
-
-[program:nova-consoleauth]
-command=/usr/bin/nova-consoleauth
-autorestart=true
-priority=30
-
-[program:nova-scheduler]
-command=/usr/bin/nova-scheduler
-autorestart=true
-priority=30
-
-[program:nova-conductor]
-command=/usr/bin/nova-conductor
-autorestart=true
-priority=30
-
-[program:nova-novncproxy]
-command=/usr/bin/nova-novncproxy
-autorestart=true
-priority=30
