--- dockerfiles/rabbitmq/supervisord.conf
@@ -1,10 +1,0 @@
-[supervisord]
-nodaemon=true
-
-[program:sshd]
-command=/usr/sbin/sshd -D
-
-[program:rabbitmq-server]
-command=/usr/sbin/rabbitmq-server
-autorestart=true
-priority=30
