+++ setup.py
@@ -14,52 +14,30 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
-import os
 from setuptools import setup, find_packages
 
 requirements = ['docker-py',
                 'PyYAML>=3.1.0',
                 'argparse',
+                'panama-template'
                 ]
                 
 testrequirements = ['nose']
 
-def get_config_files():
-
-    config_files = [('/var/lib/octopenstack', ['conf/configuration.yml',
-                                           'conf/services.yml'])]
-
-    for dirname, dirnames, filenames in os.walk('dockerfiles'):
-        for subdirname in dirnames:
-            config_dest_path = ('/var/lib/octopenstack/%s' % subdirname)
-            for dirname, dirnames, filenames in os.walk('dockerfiles/%s' % subdirname):
-                config_dir=[]
-                for filename in filenames:
-                    config_src_path = ('%s/%s' % (dirname, filename))
-                    config_dir.append(config_src_path)
-
-            config_file_dir_liste = config_dest_path, config_dir
-            config_files.append(config_file_dir_liste)
-
-
-    return config_files
-
-
 setup(
-    name='octopenstack',
+    name='panama',
     description='Easily deploy an OpenStack platform in Docker Containers',
     author='Thibaut Lapierre',
     author_email='root@epheo.eu',
-    url='https://github.com/Epheo/octopenstack',
-    # download_url='https://github.com/Epheo/octopenstack/archive/master.zip',
+    url='https://github.com/Epheo/panama',
+    # download_url='https://github.com/Epheo/panama/archive/master.zip',
     # setup_requires=['setuptools-markdown'],
     long_description_markdown_filename='README.md',
     license='Apache Software License',
     version='2015-0.5-dev',
     entry_points={'console_scripts': [
-            'octopenstack = octopenstack:__main__'
+            'panama = panama:__main__'
         ]},
-    data_files=get_config_files(),
     packages=find_packages(),
     classifiers=[
         'Development Status :: 2 - Pre-Alpha',
